Finish typing

+++ frontend/src/components/incomes/incomes.ts
@@ -1,7 +1,19 @@
 import {ModalManager} from "../modal";
 import {OperationsService} from "../../services/operations-service";
+import {
+    CategoriesResponseType,
+    CategoryErrorResponse,
+    CategorySuccessResponse
+} from "../../types/categories-response.type";
+import {OperationsResponseType, OperationsSuccessResponse} from "../../types/operations-response.type";
+import {ParamsType} from "../../types/params.type";
 
 export class Incomes {
+    readonly buttonNoDeleteElement: HTMLElement | null;
+    readonly buttonDeleteElement: HTMLElement | null;
+    readonly modalOverlay: HTMLElement | null;
+    private params: ParamsType | null;
+
     constructor() {
         this.buttonNoDeleteElement = document.getElementById("no-delete");
         this.buttonDeleteElement = document.getElementById("modal-delete");
@@ -19,13 +31,12 @@ export class Incomes {
         this.params = null;
     }
 
-    async getIncomes(params) {
+    private async getIncomes(params: string): Promise<void> {
         try {
-            const incomesResult = await OperationsService.getCategories(`/${params}`);
-            if (incomesResult && incomesResult.length > 0) {
-                this.showIncomes(incomesResult);
-            } else if (incomesResult.error) {
-                console.log(incomesResult.error);
+            const incomesResult: CategoriesResponseType = await OperationsService.getCategories(`/${params}`);
+            if (incomesResult && (incomesResult as CategorySuccessResponse[]).length > 0) {
+                this.showIncomes(incomesResult as CategorySuccessResponse[]);
+            } else {
                 location.href = '#/operations';
             }
         } catch (error) {
@@ -33,35 +44,34 @@ export class Incomes {
         }
     }
 
-    async deleteCategoryIncome() {
-        this.params = JSON.parse(this.modalOverlay.dataset.params); // получаем id и название категории, которую надо удалить
+    private async deleteCategoryIncome(): Promise<void> {
+        if (this.modalOverlay && this.modalOverlay.dataset.params) {
+            this.params = JSON.parse(this.modalOverlay.dataset.params); // получаем id и название категории, которую надо удалить
+        }
         try {
-            const operationsResult = await OperationsService.getOperations(`?period=all`); // получаем все операции для последующего удаления совпадающих с удаляемой категорией
+            const operationsResult: OperationsResponseType = await OperationsService.getOperations(`?period=all`); // получаем все операции для последующего удаления совпадающих с удаляемой категорией
             if (operationsResult) {
-                const operationsToDeleteResult = operationsResult.filter(item => item.category === this.params.category); // находим все записи, связанные с удаляемой категорией
-                if (operationsToDeleteResult) {
-                    const deleteCategoryResult = await OperationsService.deleteCategory(`/income/${this.params.id}`); // удаляем категорию
-                    if (deleteCategoryResult) {
-                        for (const item of operationsToDeleteResult) {
-                            try {
-                                const deleteOperationResult = await OperationsService.deleteOperation(`/${item.id}`); // удаляем записи, связанные с удаленной категорией
-                                if (deleteOperationResult) {
-                                    this.flagDelete = true;
+                if (this.params && 'category' in this.params && 'id' in this.params) {
+                    const operationsToDeleteResult: OperationsSuccessResponse[] = (operationsResult as OperationsSuccessResponse[]).filter((item: OperationsSuccessResponse): boolean => item.category === (this.params as ParamsType).category); // находим все записи, связанные с удаляемой категорией
+                    if (operationsToDeleteResult) {
+                        const deleteCategoryResult: CategoryErrorResponse | false = await OperationsService.deleteCategory(`/income/${(this.params as ParamsType).id}`); // удаляем категорию
+                        if (deleteCategoryResult) {
+                            for (const item of operationsToDeleteResult) {
+                                try {
+                                    await OperationsService.deleteOperation(`/${item.id}`); // удаляем записи, связанные с удаленной категорией
+                                } catch (error) {
+                                    console.log(error);
                                 }
-                            } catch (error) {
-                                console.log(error);
                             }
+                            ModalManager.hideModal();
+                            location.href = '#/incomes';
+                            console.log('DELETE CATEGORY');
+                        } else {
+                            location.href = '#/operations';
                         }
-                        ModalManager.hideModal();
-                        location.href = '#/incomes';
-                        console.log('DELETE CATEGORY');
-                    } else if (deleteCategoryResult.error) {
-                        console.log(deleteCategoryResult.error);
-                        location.href = '#/operations';
                     }
                 }
-            } else if (operationsResult.error) {
-                console.log(operationsResult.error);
+            } else {
                 location.href = '#/';
             }
         } catch (error) {
@@ -69,45 +79,47 @@ export class Incomes {
         }
     }
 
-    showIncomes(incomes) {
-        const incomesContainerElement = document.getElementById("incomes-container");
-        const blockAddingElement = document.getElementById("block-adding");
-        incomesContainerElement.innerHTML = '';
-        incomesContainerElement.appendChild(blockAddingElement);
-        incomes.forEach((income) => {
-            const div = this.createIncomeBlock(income.id, income.title);
-            incomesContainerElement.insertBefore(div, blockAddingElement);
-        });
+    private showIncomes(incomes: CategorySuccessResponse[]): void {
+        const incomesContainerElement: HTMLElement | null = document.getElementById("incomes-container");
+        const blockAddingElement: HTMLElement | null = document.getElementById("block-adding");
+        if (incomesContainerElement && blockAddingElement) {
+            incomesContainerElement.innerHTML = '';
+            incomesContainerElement.appendChild(blockAddingElement);
+            incomes.forEach((income: CategorySuccessResponse):void => {
+                const div: HTMLElement = this.createIncomeBlock(income.id.toString(), income.title);
+                incomesContainerElement.insertBefore(div, blockAddingElement);
+            });
+        }
     }
 
-    createIncomeBlock(id, title, editHref = '#/income/edit') {
+    private createIncomeBlock(id: string, title: string, editHref: string = '#/income/edit'): HTMLElement {
         // Создаем основной контейнер
-        const block = document.createElement('div');
+        const block: HTMLElement = document.createElement('div');
         block.className = 'income-block border bg-border-custom rounded';
         block.setAttribute('data-id', id);
 
         // Создаем заголовок блока
-        const titleDiv = document.createElement('div');
+        const titleDiv: HTMLElement = document.createElement('div');
         titleDiv.className = 'income-title';
 
-        const titleText = document.createElement('h3');
+        const titleText: HTMLElement = document.createElement('h3');
         titleText.className = 'income-title-text';
         titleText.textContent = title;
 
         titleDiv.appendChild(titleText);
 
         // Создаем контейнер для кнопок
-        const activeDiv = document.createElement('div');
+        const activeDiv: HTMLElement = document.createElement('div');
         activeDiv.className = 'income-active d-flex';
 
         // Кнопка "Редактировать"
-        const editButton = document.createElement('a');
+        const editButton: HTMLAnchorElement = document.createElement('a');
         editButton.href = `${editHref}?id=${id}`;
         editButton.className = 'income-edit btn btn-primary btn-custom';
         editButton.textContent = 'Редактировать';
 
         // Кнопка "Удалить"
-        const deleteButton = document.createElement('button');
+        const deleteButton: HTMLButtonElement = document.createElement('button');
         deleteButton.type = 'button';
         deleteButton.className = 'income-delete btn btn-danger btn-custom';
         deleteButton.textContent = 'Удалить';
